# Conflicts:
#	api/src/main/java/uk/ac/leeds/comp2913/api/Controller/ReceiptController.java
#	api/src/main/java/uk/ac/leeds/comp2913/api/Domain/Model/Receipt.java
#	api/src/main/resources/db/migration/V1_21__ALTER_BOOKING_TABLE_FOREIGN_RECEIPT_ID.sql

+++ api/src/main/resources/db/migration/V1_19__CREATE_USER_TABLE.sql
@@ -1,11 +1,11 @@
 create table sports_centre_management.user
 (
-    id            int8         not null,
+    id            bigserial not null,
     created_at    timestamp,
-    email_address varchar(255) not null,
-    first_name    varchar(255) not null,
-    last_name     varchar(255) not null,
-    password      varchar(255) not null,
+    email_address varchar(255)          not null,
+    first_name    varchar(255)          not null,
+    last_name     varchar(255)          not null,
+    password      varchar(255)          not null,
     updated_at    timestamp,
     primary key (id)
 );
