+++ api/src/main/resources/db/migration/V1_3__CREATE_Resource_TABLE.sql
@@ -1,6 +1,6 @@
 create table if not exists sports_centre_management.resource
 (
-    id         int8 not null,
+    id         bigserial not null,
     created_at timestamp,
     name       varchar(255),
     updated_at timestamp,
